+++ serializer/evaluation_table.go
@@ -6,16 +6,16 @@ import "PingLeMe-Backend/model"
 
 // EvaluationTable 评审表序列化器
 type EvaluationTable struct {
-	TableID	   uint					 `json:"table_id"`
+	TableID    uint                  `json:"table_id"`
 	TableName  string                `json:"table_name"`
 	TableItems []EvaluationTableItem `json:"table_items"`
 }
 
 // EvaluationTableItem 评审表项序列化器
 type EvaluationTableItem struct {
-	ItemID			uint				  `json:"item_id"`
+	ItemID          uint                  `json:"item_id"`
 	Content         string                `json:"content"`
-	Score           int                   `json:"score"`
+	Score           float32               `json:"score"`
 	Description     string                `json:"description"`
 	ChildTableItems []EvaluationTableItem `json:"child_table_items"`
 }
@@ -26,7 +26,7 @@ func BuildEvaluationTable(tableModel model.EvaluationTable) EvaluationTable {
 	table.TableName = tableModel.TableName
 	items := BuildTableItems(0, len(tableModel.TableItem)-1, tableModel.TableItem)
 	return EvaluationTable{
-		TableID: tableModel.ID,
+		TableID:    tableModel.ID,
 		TableName:  tableModel.TableName,
 		TableItems: items,
 	}
@@ -51,7 +51,7 @@ func BuildTableItems(begin, end int, tableItems []model.EvaluationTableItem) []E
 		if tableItems[i].Level == level {
 			if b != -1 {
 				items = append([]EvaluationTableItem{{
-					ItemID: 		 tableItems[i].ID,
+					ItemID:          tableItems[i].ID,
 					Content:         tableItems[i].Content,
 					Score:           tableItems[i].Score,
 					ChildTableItems: BuildTableItems(b, e, tableItems),
